+++ Source/OxyPlot.Wpf/PlotView.cs
@@ -21,6 +21,13 @@ namespace OxyPlot.Wpf
     public partial class PlotView : PlotViewBase
     {
         /// <summary>
+        /// Identifies the <see cref="TextMeasurementMethod"/> dependency property.
+        /// </summary>
+        public static readonly DependencyProperty TextMeasurementMethodProperty =
+            DependencyProperty.Register(
+                nameof(TextMeasurementMethod), typeof(TextMeasurementMethod), typeof(PlotViewBase), new PropertyMetadata(TextMeasurementMethod.TextBlock));
+
+        /// <summary>
         /// Initializes a new instance of the <see cref="PlotView" /> class.
         /// </summary>
         public PlotView()
@@ -36,9 +43,19 @@ namespace OxyPlot.Wpf
         public bool DisconnectCanvasWhileUpdating { get; set; }
 
         /// <summary>
+        /// Gets or sets the vertical zoom cursor.
+        /// </summary>
+        /// <value>The zoom vertical cursor.</value>
+        public TextMeasurementMethod TextMeasurementMethod
+        {
+            get => (TextMeasurementMethod)this.GetValue(TextMeasurementMethodProperty);
+            set => this.SetValue(TextMeasurementMethodProperty, value);
+        }
+
+        /// <summary>
         /// Gets the Canvas.
         /// </summary>
-        private Canvas Canvas => (Canvas)this.plotPresenter;
+        protected Canvas Canvas => (Canvas)this.plotPresenter;
 
         /// <summary>
         /// Gets the CanvasRenderContext.
@@ -82,7 +99,7 @@ namespace OxyPlot.Wpf
         /// <inheritdoc/>
         protected override void RenderOverride()
         {
-            RenderContext.TextMeasurementMethod = TextMeasurementMethod;
+            this.RenderContext.TextMeasurementMethod = this.TextMeasurementMethod;
             if (this.DisconnectCanvasWhileUpdating)
             {
                 // TODO: profile... not sure if this makes any difference
@@ -125,23 +142,6 @@ namespace OxyPlot.Wpf
             var exporter = new PngExporter() { Width = (int)this.ActualWidth, Height = (int)this.ActualHeight };
             var bitmap = exporter.ExportToBitmap(this.ActualModel);
             Clipboard.SetImage(bitmap);
-        }
-
-        /// <summary>
-        /// Identifies the <see cref="TextMeasurementMethod"/> dependency property.
-        /// </summary>
-        public static readonly DependencyProperty TextMeasurementMethodProperty =
-            DependencyProperty.Register(
-                nameof(TextMeasurementMethod), typeof(TextMeasurementMethod), typeof(PlotViewBase), new PropertyMetadata(TextMeasurementMethod.TextBlock));
-
-        /// <summary>
-        /// Gets or sets the vertical zoom cursor.
-        /// </summary>
-        /// <value>The zoom vertical cursor.</value>
-        public TextMeasurementMethod TextMeasurementMethod
-        {
-            get => (TextMeasurementMethod)this.GetValue(TextMeasurementMethodProperty);
-            set => this.SetValue(TextMeasurementMethodProperty, value);
         }
     }
 }